Support --type flag in update command
Adds the `--type` flag same as the `create` command to the `update`
command.
